fix(day06): ui completion and reputation logic

+++ day06/life_pro_tips/tips/views.py
@@ -5,6 +5,8 @@ from .forms import RegisterForm, LoginForm,  TipForm
 import time
 from .models import CustomUser, Tip
 from django.contrib.auth import login as auth_login, logout as auth_logout
+from django.db.models import F
+from django.db import transaction
 
 
 def home(request):
@@ -25,9 +27,7 @@ def home(request):
 
 
 def addtip(request):
-    print("USER AUTHENTICATED")
     if request.user.is_authenticated:
-        print("USER AUTHENTICATED")
         tipform = TipForm(request.POST)
         if tipform.is_valid():
             tip = tipform.save(commit=False)
@@ -37,17 +37,21 @@ def addtip(request):
 
 def deltip(request):
     if request.user.is_authenticated:
-        tip = Tip.objects.filter(id=request.POST.get("tip_id")).first()
-        if tip.author == request.user or request.user.has_perm("tips.delete_tip") or request.user.can_delete_tips() :
-            author = tip.author
-            author.reputation -= (tip.upvotes.count() * 5)
-            author.reputation += (tip.downvotes.count() * 2)
-            author.save()
-            tip.delete()
+        tip = Tip.objects.select_related('author').filter(id=request.POST.get("tip_id")).first()
+        if tip:
+            if tip.author == request.user or request.user.has_perm("tips.delete_tip") or request.user.can_delete_tips() :
+                with transaction.atomic():
+                    up_loss = tip.upvotes.count() * 5
+                    down_gain = tip.downvotes.count() * 2
+                    tip.author.reputation = F('reputation') - up_loss + down_gain
+                    tip.author.save()
+                    tip.delete()
     return (redirect('/'))
 
 
 def login(request):
+    if request.user.is_authenticated:
+        return redirect('/')
     if request.method == "POST":
         form = LoginForm(request=request, data=request.POST)
         if form.is_valid():
@@ -62,43 +66,45 @@ def login(request):
 def downvotetip(request):
     if request.user.is_authenticated:
         user = request.user
-        tip = Tip.objects.filter(id=request.POST.get("tip_id")).first()
+        tip = Tip.objects.select_related("author").filter(id=request.POST.get("tip_id")).first()
         if tip and (request.user.has_perm("tips.candownvote") or user == tip.author or user.can_downvote()):
             if user in tip.upvotes.all():
                 tip.upvotes.remove(user)
-                tip.author.reputation -= 5
+                tip.author.reputation = F('reputation') - 5
                 tip.author.save()
             if user in tip.downvotes.all():
                 tip.downvotes.remove(user)
-                tip.author.reputation += 2
+                tip.author.reputation = F('reputation') + 2
                 tip.author.save()
             else:
-                tip.author.reputation -= 2
+                tip.author.reputation = F('reputation') - 2
                 tip.downvotes.add(user)
                 tip.author.save()
     return (redirect('/'))
 
 def upvotetip(request):
     if request.user.is_authenticated:
-        tip = Tip.objects.filter(id=request.POST.get("tip_id")).first()
+        tip = Tip.objects.select_related("author").filter(id=request.POST.get("tip_id")).first()
         if tip:
             user = request.user
             if user in tip.downvotes.all():
                 tip.downvotes.remove(user)
-                tip.author.reputation += 2
+                tip.author.reputation = F('reputation') + 2
                 tip.author.save()
             if user in tip.upvotes.all():
                 tip.upvotes.remove(user)
-                tip.author.reputation -= 5
+                tip.author.reputation = F('reputation') - 5
                 tip.author.save()
             else:
-                tip.author.reputation += 5
+                tip.author.reputation = F('reputation') + 5
                 tip.upvotes.add(user)
                 tip.author.save()
     return (redirect('/'))
 
 
 def registration(request):
+    if request.user.is_authenticated:
+        return redirect('/')
     if request.method == "POST":
         form = RegisterForm(request.POST)
         if form.is_valid():
